+++ src/Persistence/Data/LocalizableString.cs
@@ -1,0 +1,8 @@
+﻿namespace PipServices.Azure.Persistence.Data
+{
+    public class LocalizableString
+    {
+        public string LocalizedValue { get; set; }
+        public string Value { get; set; }
+    }
+}