+++ src/Persistence/Data/MetricInterval.cs
@@ -1,0 +1,10 @@
+﻿namespace PipServices.Azure.Persistence.Data
+{
+    public enum MetricInterval
+    {
+        PT1M,
+        PT5M,
+        PT1H,
+        P1D
+    }
+}