+++ src/Persistence/ICosmosDbRestClient.cs
@@ -7,6 +7,7 @@ namespace PipServices.Azure.Persistence
     {
         Task<bool> CollectionExistsAsync(string correlationId);
         Task CreatePartitionCollectionAsync(string correlationId, int throughput, List<string> indexes);
+        Task<int> GetThroughputAsync(string correlationId);
         Task<bool> UpdateThroughputAsync(string correlationId, int throughput);
         Task UpdatePartitionCollectionAsync(string correlationId, List<string> indexes);
     }