+++ src/Persistence/Lock/ICosmosDbThroughputMonitorLock.cs
@@ -1,8 +1,0 @@
-﻿namespace PipServices.Azure.Persistence.Lock
-{
-    public interface ICosmosDbThroughputMonitorLock
-    {
-        void ReleaseLock(string correlationId);
-        bool TryAcquireLock(string correlationId);
-    }
-}